refactor models to enforce interface
